Add speech to text ability to text client
